Updated structure and moved constants
