+++ paraccompiler/__init__.py
@@ -12,6 +12,15 @@ __code_name__ = ""
 __release__ = f"{__code_name__} {__version__}"
 __copyright__ = "Luna Klatzer"
 
+import sys
+import pathlib
+
+# argv[0] returns the path of the starting script
+RUN_SCRIPT_DIR: pathlib.Path = pathlib.Path(__file__).parent.parent.resolve()
+
+# Adding the working directory (location of compiler-cli.py)
+sys.path.append(str(RUN_SCRIPT_DIR))
+
 # Main imports
 from . import logging
 from . import utils
@@ -21,11 +30,6 @@ from . import core
 # Importing the files that are used in the Pre-Processor and Compiler
 from .core import abc  # Abstract Classes
 
-# Preprocessor Import
-import sys
-import os
-# Adding the working directory (location of compiler-cli.py)
-sys.path.append(os.getcwd())
 # Importing from the added path the preprocessor
 import preprocessor
 
@@ -50,7 +54,6 @@ __all__ = [
     '__release__',
     '__copyright__',
     'abc',
-    'WIN',
     'preprocessor',
     'utils',
     'logging',
@@ -58,13 +61,43 @@ __all__ = [
     *core.__all__,
     *para_exceptions.__all__,
     *__main__.__all__,
+    'WIN',
+    'CONFIG_PATH',
+    'RUN_SCRIPT_DIR',
+    'VALID_FILE_ENDINGS',
+    'DEFAULT_LOG_PATH',
+    'DEFAULT_DIST_PATH',
+    'DEFAULT_BUILD_PATH',
+    'DEFAULT_CONFIG',
+    'INIT_OVERWRITE'
 ]
 
 import logging as lib_logging
 import click
 import colorama
+from typing import List
 
 
 lib_logging.getLogger(__name__).addHandler(lib_logging.NullHandler())
 colorama.init(autoreset=True)
+
 WIN: bool = click.utils.WIN
+SEPARATOR: str = "\\" if WIN else "/"
+CONFIG_PATH: str = f"{RUN_SCRIPT_DIR}{SEPARATOR}compile-config.json"
+VALID_FILE_ENDINGS: List[str] = [".para", ".parah", ".c", ".h", ".ph"]
+DEFAULT_LOG_PATH: str = "./para.log"
+DEFAULT_BUILD_PATH: str = "./build"
+DEFAULT_DIST_PATH: str = "./dist"
+# If the init overwrite is true =>
+# Existence check for the c-compiler will always return True
+INIT_OVERWRITE: bool = False
+DEFAULT_CONFIG: dict = {
+    "c-compiler-path": ""
+}
+INVALID_UNIX_FILE_NAME_CHARS: List[str] = ['/', '<', '>', '\0', '|', ':', '&']
+INVALID_WIN_FILE_NAME_CHARS: List[str] = [
+    '<', '>', ':', '"', '/', '\\', '|', '?', '*'
+]
+
+# Main-Class
+ParacCompiler = core.compiler.ParacCompiler
