+++ paraccompiler/core/__init__.py
@@ -6,6 +6,8 @@ from . import abc
 from . import error_handler
 from . import logic_stream
 from . import ctx
+from .process import *
+from . import process
 from .ctx import *
 from . import compiler
 from .compiler import *
@@ -14,6 +16,7 @@ from .. import *
 __all__ = [
     *compiler.__all__,
     *ctx.__all__,
+    *process.__all__,
     'optimiser',
     'parser',
     'abc',
